set up testing in server, restructure tests folder in src and server
