+++ server/package.json
@@ -17,7 +17,27 @@
     "redis": "^3.0.2"
   },
   "devDependencies": {
+    "@babel/core": "^7.12.13",
+    "@babel/preset-env": "^7.12.13",
+    "babel-jest": "^26.6.3",
     "jest": "^26.6.3",
     "supertest": "^6.1.3"
+  },
+  "jest": {
+    "setupFiles": [
+      "raf/polyfill"
+    ],
+    "coverageThreshold": {
+      "global": {
+        "statements": 80,
+        "branches": 80,
+        "functions": 80,
+        "lines": 80
+      }
+    },
+    "coverageReporters": [
+      "text-summary",
+      "text"
+    ]
   }
 }
